Arel::Nodes::BetweenSymmetric
